LayoutTests/fast/page-color-sampling/blend-modal-background-color.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=100 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            font-family: system-ui;
            margin: 0;
        }

        div.dimming {
            position: fixed;
            top: 0;
            left: 0;
            right: 0;
            bottom: 0;
            background: rgba(0, 0, 0, 0.6);
        }

        .tall {
            width: 100%;
            height: 2000px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(100, 0, 0, 0);
        await UIHelper.ensurePresentationUpdate();
        colors = await UIHelper.fixedContainerEdgeColors();

        shouldBeEqualToString("colors.top", "rgb(102, 102, 102)");
        finishJSTest();
    });
    </script>
</head>
<body>
<div class="dimming"></div>
<div class="tall"></div>
<header></header>
</body>
</html>
