LayoutTests/fast/page-color-sampling/blend-modal-background-color-expected.txt

PASS colors.top is "rgb(102, 102, 102)"
PASS successfullyParsed is true

TEST COMPLETE
